Update esp-nimble and pioarduino 3.1.3, various improvements and fixes for HA, SPIFFS and TOTP, add self-signed SSL certificate generator, add bypass MFA, add admin key, add build for GL-S10

+++ src/NukiOpenerWrapper.h
@@ -108,7 +108,7 @@ private:
     int _retryConfigCount = 0;
     int _retryLockstateCount = 0;
     int64_t _nextRetryTs = 0;
-    int64_t _invalidCount = 0;
+    int _invalidCount = 0;
     int64_t _lastCodeCheck = 0;
     std::vector<uint16_t> _keypadCodeIds;
     std::vector<uint32_t> _keypadCodes;
